Adds ActionCellRenderer
This adds a new CellRenderer that creates the buttons which then drive task actions.  These buttons are created with Bootstrap glyph things and the events are wired up on the Task side based on class names.
Also cleaned up how the rawData object is dealt with for the rendering with indent and now completion.